Add `feather_spi` as SPI on OrangeCrab
**Problems**

OrangeCrab has Feather compatible SPI pins but SPI is not enabled.

**Solution**

Add `feathre_spi` pads and add SPI to SoC.
This patch requires litex-hub/litex-boards#251.

NOTE: Somehow with `spisdcard`, `spi` is not working with a
distributed prebuild Linux Kernel image.
It is required to build the latest Kernel from the source to make
them work.